Move SpatialNode and ClipNode to their own files
These nodes types are not really related so it makes sense to give them
their own source files. 'clip_scroll_node' as a concept doesn't exist
any longer.
